refactor(cache): separate legacy cache implementation
