+++ .agents/CACHE_AND_INCREMENTAL.md
@@ -98,10 +98,10 @@ does not create a third public storage mode.
 
 Rspack currently carries two implementations of the same Cache responsibility:
 
-| Backend        | Location                            |
-| -------------- | ----------------------------------- |
-| `legacy_cache` | `crates/rspack_core/src/cache/`     |
-| `new_cache`    | `crates/rspack_core/src/new_cache/` |
+| Backend        | Location                                  |
+| -------------- | ----------------------------------------- |
+| `legacy_cache` | `crates/rspack_core/src/legacy_cache/`   |
+| `new_cache`    | `crates/rspack_core/src/new_cache/`      |
 
 The backend selector is not another Incremental mode. Switching between `legacy_cache` and
 `new_cache` may change how cache entries are keyed, retained, serialized, or flushed, but it must not
@@ -168,7 +168,8 @@ introduce new Cache/Incremental coupling.
 ## Relevant Code
 
 - Cache configuration: `crates/rspack_core/src/options/cache.rs`
-- Legacy Cache backend: `crates/rspack_core/src/cache/`
+- Shared Cache dependencies: `crates/rspack_core/src/cache/`
+- Legacy Cache backend: `crates/rspack_core/src/legacy_cache/`
 - New Cache backend: `crates/rspack_core/src/new_cache/`
 - Incremental configuration and mutations: `crates/rspack_core/src/incremental/`
 - Incremental artifact ownership: `crates/rspack_core/src/artifacts/incremental_artifacts.rs`
