+++ crates/rspack_core/src/cache/build_dependencies/utils.rs
