+++ crates/rspack_core/src/cache/build_dependencies/visitor.rs
