+++ crates/rspack_core/src/cache/codec.rs
