+++ crates/rspack_core/src/cache/snapshot.rs
