+++ crates/rspack_core/src/compilation/build_module_graph/pass.rs
@@ -4,11 +4,11 @@ use rspack_error::Result;
 use super::build_module_graph_pass;
 use crate::{
   Compilation,
-  cache::Cache,
   compilation::{
     finish_make::finish_make_pass, finish_module_graph::finish_module_graph_pass,
     make::make_hook_pass, pass::PassExt,
   },
+  legacy_cache::Cache,
   logger::Logger,
 };
