+++ crates/rspack_core/src/compiler/mod.rs
@@ -22,10 +22,10 @@ use crate::{
   CompilerPlatform, ContextModuleFactory, Filename, KeepPattern, NormalModuleFactory, PluginDriver,
   ResolverFactory, SharedPluginDriver,
   artifacts::IncrementalArtifacts,
-  cache::{Cache as LegacyCache, new_cache as create_legacy_cache},
   compilation::build_module_graph::ModuleExecutor,
   fast_set, include_hash,
   incremental::{Incremental, IncrementalPasses},
+  legacy_cache::{Cache as LegacyCache, create_cache as create_legacy_cache},
   logger::Logger,
   new_cache::{Cache, CacheFacade, create_cache},
   trim_dir,
