+++ crates/rspack_core/src/legacy_cache/disable.rs
