+++ crates/rspack_core/src/legacy_cache/memory.rs
