+++ crates/rspack_core/src/legacy_cache/mixed.rs
