+++ crates/rspack_core/src/legacy_cache/persistent/context.rs
