+++ crates/rspack_core/src/legacy_cache/persistent/occasion/make/alternatives/dependency.rs
