+++ crates/rspack_core/src/legacy_cache/persistent/snapshot/scope.rs
