+++ crates/rspack_core/src/legacy_cache/persistent/snapshot/strategy/package_helper.rs
