+++ crates/rspack_core/src/legacy_cache/persistent/storage/mod.rs
@@ -1,22 +1,10 @@
 use std::sync::Arc;
 
-use rspack_cacheable::{cacheable, utils::PortablePath, with::As};
 use rspack_fs::IntermediateFileSystem;
-use rspack_paths::Utf8PathBuf;
 pub use rspack_storage::{BoxStorage, CacheDirectory, MemoryStorage, Storage};
 use rspack_storage::{FileSystemOptions, FileSystemStorage};
 
-/// Storage Options
-///
-/// This enum contains all of supported storage options.
-#[cacheable]
-#[derive(Debug, Clone, Hash)]
-pub enum StorageOptions {
-  FileSystem {
-    #[cacheable(with=As<PortablePath>)]
-    directory: Utf8PathBuf,
-  },
-}
+use crate::cache::StorageOptions;
 
 pub fn create_storage(
   options: StorageOptions,
